\name{CutPointTests}
\alias{CutPointTests}
\title{Computes odds ratio and p-value of interaction term of variable and treatment}
\usage{
CutPointTests(c, response, var, treat, covars, data)
}
\arguments{
  \item{c}{cutpoint for variable}

  \item{response}{the response, see
  \code{link{OptimalCutPoint}}

  \item{var}{the variable, see \code{link{OptimalCutPoint}}

  \item{treat}{treatment, see \code{link{OptimalCutPoint}}

  \item{covars}{covariates, see
  \code{link{OptimalCutPoint}}

  \item{data}{see \code{link{OptimalCutPoint}}
}
\description{
Helper function for parallel processing
}

