\name{SplitVec}
\alias{SplitVec}
\title{function to split a vector in pieces, needed for e.g. cross validation}
\usage{
SplitVec(x, n)
}
\arguments{
  \item{x}{the vector to split}

  \item{n}{the number of splits}
}
\description{
function to split a vector in pieces, needed for e.g. cross
validation
}

