\name{STEPPBoxPlot}
\alias{STEPPBoxPlot}
\title{Helper function to produce a boxplot of the continuous variable}
\usage{
STEPPBoxPlot(var, ylab)
}
\arguments{
  \item{var}{the continuous variable as numeric vector}

  \item{ylab}{the label for the y axis}
}
\value{
a ggplot object with the boxplot
}
\description{
Helper function to produce a boxplot of the continuous
variable
}

