\name{STEPPDensityPlot}
\alias{STEPPDensityPlot}
\title{Helper function to produce a density plot of the continuous variable}
\usage{
STEPPDensityPlot(var, xlab)
}
\arguments{
  \item{var}{the continuous variable as numeric vector}

  \item{xlab}{the label for the x axis}
}
\value{
a ggplot object with the density plot
}
\description{
Helper function to produce a density plot of the continuous
variable
}

