\name{STEPPOmnibusTest}
\alias{STEPPOmnibusTest}
\title{Helper function to peform an Omnibus test for the equality of coefficients across subpopulations (groups)}
\usage{
STEPPOmnibusTest(coefs, treat, groups)
}
\arguments{
  \item{coefs}{matrix with coefficients and variances}

  \item{treat}{the treatment variable as two level factor}

  \item{groups}{the list with groups}
}
\value{
list with
}
\description{
Helper function to peform an Omnibus test for the equality
of coefficients across subpopulations (groups)
}

