\name{STEPPTail}
\alias{STEPPTail}
\title{Helper function to get a tail-oriented (both parts) subpopulation pattern}
\usage{
STEPPTail(x, k)
}
\arguments{
  \item{x}{vector with covariate}

  \item{k}{number of groups. Should be odd (2*g+1), if even
  (k+1) is used}
}
\description{
Helper function to get a tail-oriented (both parts)
subpopulation pattern
}

